UT Football Twitter Sentiment Analysis
How often have you been watching something pivotal on TV (big football
game? news?) and think to yourself, “wow, I wonder what the rest of the
world
is thinking about this right now?” If you’re anything like me, it’s
relatively often. It happened to me last night as I sat at my desk
working, listening to my wife shouting for her football team, UT, in
their last game this season against Baylor. And this was a BIG game.

Eventually it occurred to me that I CAN, in fact, quantify how the
world is reacting. I threw together a quick script to query Twitter for
the hashtag, “#hookem” (the battle cry of Longhorns). The truly
interesting part of this query was that I confined it to geo-enabled
tweets between the SW corner of the UT campus and the NE corner of the
UT campus.  In other words, these were only UT students’ sentiments on
the game (hypothetically), so I figured they would be most charged.
Unfortunately, I didn’t think to do it until the 4th quarter, but it
was enough to work with — I also anticipated I’d catch some of the
post-game aftermath, be it very positively charged or very negatively
charged. As a side note, the first time you do this, it is a slight
pain to set up an OAuth between your computer and the Twitter API, but
after it is done it is a mere matter of referencing the authentication
certificate (I keep mine in the root directory of my “Twitter
Analytics” folder). For the sake of this example, I am not going into
that… maybe another time.  My hope was that by monitoring sentiments
across time, I would be able to look at the data and say, “THAT is
where Baylor intercepted the ball/took a knee.”

The quick, (highly-simplified) pseudo code is as follows: Connect to
Twitter, capture all tweets that used #hookem within the
geo-coordinates specified (3 hour process… use caution if you hope to
replicate). Dump these tweets into a folder called “Output.” Parse the
tweets, separating them into their separate components. Identify which
column of the vector contains the info in which I am interested (I
wanted text and time-posted), parse each tweet and its time into two
respective lists. At this point I wanted to write the lists out into a
text file to finish processing in Java, but decided I’d challenge
myself and keep it all in R. This proved much more difficult than
expected, as it seems people like to use non-UTF-8 characters in their
tweets, which (to my knowledge), R can’t handle. This caused a LOT of
debugging. Without going into detail about all I tried over the
following 12 hours, I eventually decided to throw out tweets with
characters that caused trouble.  All of this was executed in one script
— that is to say, I was not manually going into files, editing them or
copying them from one place to another… that is cheating!

Eventually I parsed it all down to two refined lists of tweets’
sentiments (using the polarity function from the “qdap” library) and
their corresponding times. Then I tried to plot it… turns out R 3.0.2
hates me and continually froze. Not a surprise, given I was trying to
plot some 20,000 data points. I ended up writing the lists to CSV and
plotting in Excel. This is the only point where I manually handled any
data, but only so far as to import the .csvs and plot them.

Looks like my hypothesis was WAY off! The sentiments over time were
relatively uniform.  They are modeled on the Y-Axis in normalized,
arbitrary units. Positive numbers are positive sentiments, and negative
are negative. A sentiment of 0 does not mean “some positive, some
negative, or just OK,” it means “unknown,” so there is absolutely room
for error in the data.

Make of it what you will, but sentiments were all over the place in
that three hour period. Perhaps UT is just a very moody place, or maybe
they didn’t really care about the football game at all. It could also
be that #hookem was a poor choice of hashtag and did not capture tweets
about the game; maybe all Longhorns living on campus end their tweets
in #hookem? Of course, there always exists the possibility I screwed up
my analysis, but we will assume that is a moot point. ;) As always,
please leave me feedback if you have it.

tgsmith61591@gmail.com